Final Project for UNCW Text and Unstructured Data
This repository contains the Final Course Project I created for my Text and Unstructured Data Analytics, graduate level course, at the University of North Carolina Wilmington in April 2024. The goal of this project was to analyze the entire script of the television adaption of the Song of Ice and Fire series (Game of Thrones) across each season that aired and also, the context of the words spoken within the script by the characters. The "Analysis.pdf" file contains my full analysis and my entire R code I wrote to determine my findings.